+++ fabric/src/main/java/com/more_relics/fabric/FabricMod.java
@@ -3,6 +3,7 @@ package com.more_relics.fabric;
 import com.more_relics.fabric.compat.CompatFeatures;
 import more_relics.MoreRelics;
 import more_relics.item.Group;
+import more_relics.item.MoreRelicsItems;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.minecraft.registry.Registries;
@@ -19,6 +20,7 @@ public final class FabricMod implements ModInitializer {
         Group.GROUP = FabricItemGroup.builder()
                 .icon(Group.ICON)
                 .displayName(Text.translatable(Group.translationKey))
+                .entries((ctx, entries) -> MoreRelicsItems.addToGroup(entries))
                 .build();
         Registry.register(Registries.ITEM_GROUP, Group.KEY, Group.GROUP);