+++ neoforge/src/main/java/com/more_relics/neoforge/compat/CompatFeatures.java
@@ -1,9 +1,9 @@
 package com.more_relics.neoforge.compat;
 
-import more_relics.compat.AccessoriesCompat;
+import com.more_relics.neoforge.compat.curios.CuriosCompat;
 
 public class CompatFeatures {
     public static void init() {
-        AccessoriesCompat.init();
+        CuriosCompat.init();
     }
 }