+++ neoforge/src/main/java/com/more_relics/neoforge/compat/curios/CuriosCompat.java
@@ -1,0 +1,11 @@
+package com.more_relics.neoforge.compat.curios;
+
+import net.spell_engine.Platform;
+
+public class CuriosCompat {
+    public static void init() {
+        if (Platform.util().isModLoaded("curios")) {
+            CuriosHelper.registerFactory();
+        }
+    }
+}